+++ packages/devtools_app_shared/CHANGELOG.md
@@ -6,6 +6,7 @@ found in the LICENSE file or at https://developers.google.com/open-source/licens
 ## 0.4.1 (not released)
 * Ignore expected exceptions from the `navigateToCode` extension method.
 * Add `RpcErrorExtension` extension with an `isServiceDisposedError` getter.
+* Export `package:dtd/dtd.dart` from `dtd_manager.dart`.
 
 ## 0.4.0
 * Bump `dtd` dependency to `^4.0.0`.