+++ packages/devtools_app_shared/CHANGELOG.md
@@ -3,6 +3,9 @@ Copyright 2025 The Flutter Authors
 Use of this source code is governed by a BSD-style license that can be
 found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 -->
+## 0.5.1 (not released)
+* Add DevTools-styled text field `DevToolsTextField`.
+
 ## 0.5.0
 * **Breaking change:** remove `scaleByFontFactor`.
 * **Breaking change:** remove `IdeTheme.fontSize` and `IdeTheme.fontSizeFactor`.